Implemented VeRDET change detection method.
In this implementation, it is assumed that each time only a single index is run. In the original Matlab code, both function assumes a multi-dimentional array, which is not the case in this implementation. The pre-processing steps for VeRDET was not implemented in this class: spatial and temporal filter. I am thinking those can be done with existing GEE api.