- add vct test code
Note: This test code create the exact output for the data in the test folder. However, when running against a larger 1000x1000 test data, there are 15 pixels that do not match exactly with the original implementation. We are attributing the difference to the difference in how mean and standard deviation are calculated. In the Java implementation, it is more robust to account for the floating point representation.